Support Multiple inverter api for inverter type (#96)
* Support Multiple inverter api for inverter type

The main problem is that inverters apis are different in two ways:

- The api end point to get the data
- Some aspects of parsing the non-"data" attributes of the response
  (like SN, version, etc...)

Each single inverter could expose (for example) two end points in
general. as in, before version x it exposes an end point with query
parameters, but after version x it exposes the end point with params in
the request body.

The inheritance model in the project is short handed
here. As it is based on the assumption that a single inverter will
always support a single api end point contract.

This patch is basically (and only for now), refactoring in favor of
composition (vs. inheritance).

Each inverter class is initiated with two dependencies/components:

- InverterHttpClient
- ResponseParser

To support multiple variants for each inverter, we only build the
inverter with proper variations of those dependencies.

So, After this patch, the discovery process is all about asking the
inverter class to build all possible variants this inverter knows it
supports. The discovery then goes on, as was before, trying to use that
variant to get_data, until one variant is successful.

The patch tries as much as possible to continue providing sane and most
common defaults for inveter's implementations.

At this moment, the patch does not add (or subtract) any feature/support
on top of the baseline (before the patch). This is just opening the
opportunity to easily add missing variants.

Closes #88

WIP

* Add a verify script to be run locally

This will be helpful for new contributors to check their code against
the github workflow checks without back and forth communication with the
maintainers to run the pipeline manually for them.

Co-authored-by: Robin Wohlers-Reichel <me@robinwr.com>